feat: реализовать полноценную DTO-архитектуру
- Все контроллеры теперь работают исключительно с DTO:
  - FilmController, UserController, GenreController, MpaController
  - Абстрактный контроллер типизирован по DTO
  - Принимаем и возвращаем только DTO в запросах/ответах

- Сервисный слой использует мапперы для преобразования DTO-Entity:
  - FilmServiceImpl, UserServiceImpl, GenreServiceImpl, MpaServiceImpl
  - Все методы сервисов работают с DTO на входе и выходе
  - Бизнес-логика оперирует Entity, но границы четко определены

- Добавлены и используются все необходимые мапперы:
  - FilmMapper, UserMapper, GenreMapper, MpaMapper
  - Настроена интеграция с Spring (componentModel = 'spring')

- Все эндпоинты возвращают ResponseEntity:
  - Явное указание HTTP-статусов
  - Единообразный формат ответов
  - Подготовка для будущего расширения (заголовки, кастомные статусы)

- Сохранена валидация на уровне DTO:
  - Аннотации валидации в DTO-классах
  - Бизнес-валидация в отдельных валидаторах
  - Четкое разделение ответственности между слоями

+++ src/main/java/ru/yandex/practicum/filmorate/managment/db/GenreDbStorage.java
@@ -25,7 +25,7 @@ public class GenreDbStorage {
         return jdbcTemplate.query(sql, new GenreRowMapper());
     }
 
-    public Optional<Genre> getGenreById(Integer id) {
+    public Optional<Genre> getGenreById(Long id) {
         String sql = "SELECT * FROM genres WHERE id = ?";
         log.debug("Поиск жанра по ID: {}", id);
         List<Genre> result = jdbcTemplate.query(sql, new GenreRowMapper(), id);
@@ -44,7 +44,7 @@ public class GenreDbStorage {
         @Override
         public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
             return Genre.builder()
-                    .id(rs.getInt("id"))
+                    .id(rs.getLong("id"))
                     .name(rs.getString("name"))
                     .build();
         }
